refactor: Player와 Dealer의 sealed class 생성
