--- src/main/kotlin/blackjack/domain/player/PlayerResult.kt
@@ -1,3 +1,0 @@
-package blackjack.domain.player
-
-data class PlayerResult(val player: Player, val score: Int)
