+++ src/main/kotlin/blackjack/participant/Dealer.kt
@@ -1,0 +1,5 @@
+package blackjack.participant
+
+class Dealer {
+
+}
