+++ src/main/kotlin/blackjack/participant/player/PlayerResult.kt
@@ -1,0 +1,5 @@
+package blackjack.participant.player
+
+import blackjack.participant.Player
+
+data class PlayerResult(val player: Player, val score: Int)
