+++ src/main/kotlin/blackjack/participant/player/PlayerResults.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.player
+package blackjack.participant.player
 
 @JvmInline
 value class PlayerResults(private val results: List<PlayerResult>) : List<PlayerResult> by results
